Add via_csv_to_data_csv.py
Added the VIA.csv to to data.csv utility script. It takes a csv output
file from VGG Image Annotator (VIA) and converts it to the format we need
for using Youness Mansar's Fast_Image_Classification model training code.
Updated the README to include instructions on using it.
Also renamed the scripts to be more python-like, I think? And put them
in a utils directory.